Add experimental additional source for data provided via stdin
As per unix custom, using '-' as the source prompts the mapper to
read the data from stdin. To illustrate one use for this I've also
added a new method on the launcher `launchMappingFromJsonData` which
runs the mapping on a provided JsonNode object. Behind the scenes this
using the stdin input, which is a bit of a hack tbh.

A practical use for this is pushing data into the mapper, rather than
letting it fetch data from sources, which is something you might want
to do in a streaming context.

I'm not going to pretend this is the optimal implementation however!